Title:fix, 在光驱中选择已经刻录的文件夹，拖住到空白处，出现错误提示
Description: bug-58324
[步骤]

1.打开文件管理器进入到光驱目录中，选择任意文件夹拖拽到光驱的空白处，观察现象
[结果]

1.出现错误提示

RootCause: 光盘中的目录是只读的，拖动会产生剪切操作，试图删除目录

Solution: 光盘中的目录进行删除时，拒绝该操作

+++ dde-file-manager-lib/io/dfilecopymovejob.cpp
@@ -1110,6 +1110,11 @@ bool DFileCopyMoveJobPrivate::mergeDirectory(DFileHandler *handler, const DAbstr
         return true;
     }
 
+    //  光盘中的目录不能被删除
+    if (deviceListener->isFileFromDisc(fromInfo->fileUrl().toLocalFile())) {
+        qInfo() << "remove file from disc, reject.";
+        return true;
+    }
     // 完成操作后删除原目录
     return removeFile(handler, fromInfo);
 }
